+++ lib/obsidian/parser/markdown_content.rb
@@ -2,14 +2,15 @@
 
 module Obsidian
   class MarkdownContent
-    def initialize(path, root)
+    def initialize(path, root, renderer:)
       @path = path
       @root = root
+      @renderer = renderer
     end
 
     def generate_html
       markdown = @path.read
-      Obsidian::ObsidianFlavoredMarkdown.parse(markdown, root: @root).to_html
+      Obsidian::ObsidianFlavoredMarkdown.parse(markdown, root: @root, renderer: @renderer).to_html
     end
   end
 end