+++ spec/obsidian/parser_spec.rb
@@ -39,7 +39,7 @@ RSpec.describe Obsidian::Parser do
   end
 
   it "converts markdown into HTML content" do
-    expect(parser.pages.find { |note| note.title == "cat" }.content.generate_html).to eq("<h2 id=\"cats-are-the-best\">Cats are the best</h2>\n\n<p>Meow meow meow</p>\n")
+    expect(parser.pages.find { |note| note.title == "cat" }.content.generate_html).to eq("<h2 id=\"cats-are-the-best\">Cats are the best</h2>\n<p>Meow meow meow</p>\n")
   end
 
   it "adds index.md content to index pages" do