+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+- Replace Kramdown with Markly
+- Enabled support for Github Flavored Markdown tables and tasklists
 
 ## [0.5.4] - 2023-08-02
 - Fix page getting clobbered when wikilinks point to non-existent pages.